+++ src/main/java/com/example/paymentservice/web/dto/ResponseDto.java
@@ -4,6 +4,8 @@ import com.example.paymentservice.app.enums.PaymentStatus;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import org.springframework.lang.NonNull;
 
+import java.math.BigDecimal;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ResponseDto {
     @NonNull
@@ -11,7 +13,7 @@ public class ResponseDto {
     @NonNull
     private Long orderId;
     private String date;
-    private Integer price;
+    private BigDecimal price;
 
     public ResponseDto(@NonNull PaymentStatus status, @NonNull Long orderId) {
         this.status = status;
@@ -20,7 +22,7 @@ public class ResponseDto {
         this.price = null;
     }
 
-    public ResponseDto(@NonNull PaymentStatus status, @NonNull Long orderId, String date, Integer price) {
+    public ResponseDto(@NonNull PaymentStatus status, @NonNull Long orderId, String date, BigDecimal price) {
         this.status = status;
         this.orderId = orderId;
         this.date = date;
@@ -53,11 +55,11 @@ public class ResponseDto {
         this.date = date;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
